feat: roll up git status
Right now, this will pick the first status that isn't empty
from within a directory and apply them to the top level
status of a directory. This seems like a better compromise
than just showing empty in the status column?

But maybe it doesn't make sense, not sure.

live with it for a while to see.